+++ tests/drmcp/unit/workload_tools/test_tools.py
@@ -26,10 +26,8 @@ from datarobot_genai.drmcputils.exceptions import ToolError
 from datarobot_genai.drmcputils.exceptions import ToolErrorKind
 from datarobot_genai.drtools.workload import artifact_tools
 from datarobot_genai.drtools.workload import build_tools
-from datarobot_genai.drtools.workload import credential_tools
 from datarobot_genai.drtools.workload import lifecycle_tools
 from datarobot_genai.drtools.workload import observability_tools
-from datarobot_genai.drtools.workload import openapi_tools
 from datarobot_genai.drtools.workload import proton_tools
 from datarobot_genai.drtools.workload import read_tools
 from datarobot_genai.drtools.workload import replacement_tools
@@ -1773,215 +1771,3 @@ async def test_workload_replacement_delete_not_found(patched_dr_client: MagicMoc
         await replacement_tools.workload_replacement_delete(workload_id="wkld-missing")
     assert exc_info.value.kind is ToolErrorKind.NOT_FOUND
 
-
-# ================================================================== #
-# PR8 — credentials + read_openapi_spec                               #
-# ================================================================== #
-
-# ------------------------------------------------------------------ #
-# credential_list                                                       #
-# ------------------------------------------------------------------ #
-
-
-@pytest.mark.asyncio
-async def test_credential_list_success(patched_dr_client: MagicMock) -> None:
-    patched_dr_client.get.return_value = MagicMock(
-        json=lambda: {
-            "data": [
-                {"credentialId": "cred-1", "name": "my-s3", "credentialType": "s3"},
-                {"credentialId": "cred-2", "name": "my-basic", "credentialType": "basic"},
-            ],
-            "count": 2,
-            "totalCount": 2,
-            "next": None,
-            "previous": None,
-        }
-    )
-
-    result = await credential_tools.credential_list(limit=50)
-
-    patched_dr_client.get.assert_called_once_with("credentials/", params={"limit": 50, "offset": 0})
-    assert len(result["credentials"]) == 2
-    assert result["credentials"][0]["credentialId"] == "cred-1"
-
-
-@pytest.mark.asyncio
-async def test_credential_list_negative_offset_raises() -> None:
-    with pytest.raises(ToolError) as exc_info:
-        await credential_tools.credential_list(offset=-1)
-    assert exc_info.value.kind is ToolErrorKind.VALIDATION
-
-
-@pytest.mark.asyncio
-async def test_credential_list_client_error(patched_dr_client: MagicMock) -> None:
-    patched_dr_client.get.side_effect = ClientError("403", status_code=403, json={})
-    with pytest.raises(ToolError) as exc_info:
-        await credential_tools.credential_list()
-    assert exc_info.value.kind is ToolErrorKind.UPSTREAM
-
-
-# ------------------------------------------------------------------ #
-# credential_get                                                        #
-# ------------------------------------------------------------------ #
-
-
-@pytest.mark.asyncio
-async def test_credential_get_success(patched_dr_client: MagicMock) -> None:
-    patched_dr_client.get.return_value = MagicMock(
-        json=lambda: {
-            "credentialId": "cred-1",
-            "name": "my-s3",
-            "credentialType": "s3",
-            "description": "prod S3 bucket",
-        }
-    )
-
-    result = await credential_tools.credential_get(credential_id="cred-1")
-
-    patched_dr_client.get.assert_called_once_with("credentials/cred-1/")
-    assert result["credentialType"] == "s3"
-
-
-@pytest.mark.asyncio
-async def test_credential_get_empty_id_raises() -> None:
-    with pytest.raises(ToolError) as exc_info:
-        await credential_tools.credential_get(credential_id="")
-    assert exc_info.value.kind is ToolErrorKind.VALIDATION
-
-
-@pytest.mark.asyncio
-async def test_credential_get_not_found(patched_dr_client: MagicMock) -> None:
-    patched_dr_client.get.side_effect = ClientError("404", status_code=404, json={})
-    with pytest.raises(ToolError) as exc_info:
-        await credential_tools.credential_get(credential_id="cred-missing")
-    assert exc_info.value.kind is ToolErrorKind.NOT_FOUND
-
-
-# ------------------------------------------------------------------ #
-# credential_keys                                                       #
-# ------------------------------------------------------------------ #
-
-
-@pytest.mark.asyncio
-async def test_credential_keys_basic() -> None:
-    result = await credential_tools.credential_keys(credential_type="basic")
-    assert result["credentialType"] == "basic"
-    assert "login" in result["keys"]
-    assert "password" in result["keys"]
-
-
-@pytest.mark.asyncio
-async def test_credential_keys_s3() -> None:
-    result = await credential_tools.credential_keys(credential_type="s3")
-    assert "awsAccessKeyId" in result["keys"]
-    assert "awsSecretAccessKey" in result["keys"]
-
-
-@pytest.mark.asyncio
-async def test_credential_keys_empty_type_raises() -> None:
-    with pytest.raises(ToolError) as exc_info:
-        await credential_tools.credential_keys(credential_type="")
-    assert exc_info.value.kind is ToolErrorKind.VALIDATION
-
-
-@pytest.mark.asyncio
-async def test_credential_keys_unknown_type_raises() -> None:
-    with pytest.raises(ToolError) as exc_info:
-        await credential_tools.credential_keys(credential_type="foobar")
-    assert exc_info.value.kind is ToolErrorKind.VALIDATION
-
-
-@pytest.mark.asyncio
-async def test_credential_keys_all_known_types() -> None:
-    from datarobot_genai.drmcputils.constants import CREDENTIAL_TYPE_KEYS
-
-    for ctype in CREDENTIAL_TYPE_KEYS:
-        result = await credential_tools.credential_keys(credential_type=ctype)
-        assert result["keys"], f"Empty keys for type {ctype}"
-
-
-# ------------------------------------------------------------------ #
-# read_openapi_spec                                                     #
-# ------------------------------------------------------------------ #
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_remote_success(patched_dr_client: MagicMock) -> None:
-    patched_dr_client.get.return_value = MagicMock(
-        json=lambda: {"openapi": "3.1.0", "info": {"title": "Workload API"}, "paths": {}}
-    )
-
-    result = await openapi_tools.read_openapi_spec()
-
-    patched_dr_client.get.assert_called_once_with("openapi.json")
-    assert result["openapi"] == "3.1.0"
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_custom_remote_path(patched_dr_client: MagicMock) -> None:
-    patched_dr_client.get.return_value = MagicMock(
-        json=lambda: {"openapi": "3.1.0", "paths": {"/workloads/": {}}}
-    )
-
-    result = await openapi_tools.read_openapi_spec(remote_path="workload-openapi.json")
-
-    patched_dr_client.get.assert_called_once_with("workload-openapi.json")
-    assert "/workloads/" in result["paths"]
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_local_path_not_found() -> None:
-    with pytest.raises(ToolError) as exc_info:
-        await openapi_tools.read_openapi_spec(local_path="/nonexistent/openapi.yaml")
-    assert exc_info.value.kind is ToolErrorKind.NOT_FOUND
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_local_path_json(tmp_path: Path) -> None:
-    import json as _json
-
-    spec = {"openapi": "3.1.0", "info": {"title": "Test"}, "paths": {}}
-    spec_file = tmp_path / "spec.json"
-    spec_file.write_text(_json.dumps(spec))
-
-    result = await openapi_tools.read_openapi_spec(local_path=str(spec_file))
-    assert result["info"]["title"] == "Test"
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_local_path_yaml(tmp_path: Path) -> None:
-    spec_file = tmp_path / "spec.yaml"
-    spec_file.write_text("openapi: '3.1.0'\ninfo:\n  title: YAMLTest\npaths: {}\n")
-
-    result = await openapi_tools.read_openapi_spec(local_path=str(spec_file))
-    assert result["info"]["title"] == "YAMLTest"
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_local_path_invalid_json_raises(tmp_path: Path) -> None:
-    bad_file = tmp_path / "bad.json"
-    bad_file.write_text("{ this is not valid json !!!")
-
-    with pytest.raises(ToolError) as exc_info:
-        await openapi_tools.read_openapi_spec(local_path=str(bad_file))
-    assert exc_info.value.kind is ToolErrorKind.UPSTREAM
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_remote_non_client_error_falls_through(
-    patched_dr_client: MagicMock,
-) -> None:
-    patched_dr_client.get.side_effect = ValueError("unexpected json decode error")
-    with pytest.raises(ToolError) as exc_info:
-        await openapi_tools.read_openapi_spec()
-    assert exc_info.value.kind is ToolErrorKind.NOT_FOUND
-
-
-@pytest.mark.asyncio
-async def test_read_openapi_spec_remote_error_raises(patched_dr_client: MagicMock) -> None:
-    from datarobot.errors import ClientError as _ClientError
-
-    patched_dr_client.get.side_effect = _ClientError("503", status_code=503, json={})
-    with pytest.raises(ToolError) as exc_info:
-        await openapi_tools.read_openapi_spec()
-    assert exc_info.value.kind is ToolErrorKind.NOT_FOUND
